+++ tests/integration/main.rs
@@ -152,6 +152,8 @@ mod sync_init;
 #[cfg(feature = "sync")]
 mod sync_pair;
 #[cfg(feature = "sync")]
+mod sync_setup;
+#[cfg(feature = "sync")]
 mod sync_transport;
 mod timings;
 mod track_ta12;